requirements.txt
Flask==3.0.3
scikit-learn==1.5.0
mysql-connector-python==8.4.0
numpy==2.0.0
openai==1.35.1
pandas==2.2.2
requests==2.32.3
pdfminer.six==20231228